Should show recently processed claims; cannot test because of database access issues
